requirements.txt
Django==1.8.18
django-filter==1.0.4
djangorestframework==3.6.3
Markdown==2.6.8
gunicorn==19.6.0
whitenoise==3.2.2
psycopg2==2.6.2
